rename this type to better illustrate its purpose
